chore: ignore `&pebble.Options` and move `SetupDatabase/ClearDatabase` to `internal/database`
